+++ R/ampute.R
@@ -148,7 +148,7 @@
 #'
 #' @return Returns an S3 object of class \code{\link{mads}} (multivariate
 #' amputed data set)
-#' @author Rianne Schouten [aut, cre], Gerko Vink [aut], Peter Lugtig [ctb], 2016
+#' @author Rianne Schouten, Gerko Vink, Peter Lugtig, 2016
 #' @seealso \code{\link{mads}}, \code{\link{bwplot.mads}},
 #' \code{\link{xyplot.mads}}
 #'