+++ R/ibind.R
@@ -97,6 +97,7 @@ ibind <- function(x, y) {
     predictorMatrix = x$predictorMatrix,
     visitSequence = visitSequence,
     formulas = x$formulas,
+    modeltype = x$modeltype,
     post = x$post,
     blots = x$blots,
     ignore = x$ignore,